--- modules/parser/src/test/scala/playground/smithyql/parser/ListParserTests.scala
@@ -1,10 +1,0 @@
-package playground.smithyql.parser
-
-import playground.smithyql._
-
-import Diffs._
-import Codecs._
-
-object ListParserTests extends ParserSuite {
-  loadParserTests[Listed]("listed", trimWhitespace = true)
-}
