--- modules/parser/src/test/scala/playground/smithyql/parser/QueryParserTests.scala
@@ -1,10 +1,0 @@
-package playground.smithyql.parser
-
-import playground.smithyql._
-
-import Diffs._
-import Codecs._
-
-object QueryParserTests extends ParserSuite {
-  loadParserTests[Query]("query")
-}
