--- modules/parser/src/test/scala/playground/smithyql/parser/StructParserTests.scala
@@ -1,10 +1,0 @@
-package playground.smithyql.parser
-
-import playground.smithyql._
-
-import Diffs._
-import Codecs._
-
-object StructParserTests extends ParserSuite {
-  loadParserTests[Struct]("struct", trimWhitespace = true)
-}
