+++ modules/parser/src/test/scala/playground/smithyql/parser/generative/QueryParserTests.scala
@@ -1,0 +1,10 @@
+package playground.smithyql.parser.generative
+
+import playground.smithyql.Diffs._
+import playground.smithyql._
+import playground.smithyql.parser.Codecs._
+import playground.smithyql.parser.ParserSuite
+
+object QueryParserTests extends ParserSuite {
+  loadParserTests[Query]("query")
+}
